backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.3
numpy==1.25.2
yfinance==0.2.28
google-generativeai==0.3.2